add lession 3 Spring boot validate message anotation

+++ Nguyen_Thanh_Kien/springboot/AccountManagement/src/main/java/vti/accountmanagement/repository/PositionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import vti.accountmanagement.model.Position;
 
 @Repository
-public interface PositionRepository extends JpaRepository<Position, Long> {
-    Position findByPositionId(long id);
+public interface PositionRepository extends JpaRepository<Position, Integer> {
+    Position findByPositionId(int id);
     Position findByPositionName(String name);
 }
